refactor: reorganize components and move pages into pages/
